// types.ts
export interface Product {
  id: string;
  name: string;
  price: number;
  description: string;
  category: string;
  image: string; 
  images?: string[]; 
  modelUrl?: string; // 3D model link (.glb or .gltf)
  videoUrl?: string; // Video review link
  brand?: string;
  stock: number;
  rating: number;
  numReviews?: number;
  featured?: boolean;
  isOffer?: boolean;
  offerPrice?: number;
  offerEndTime?: number;
  views?: number;
}
// ...
export interface Review {
  id: string;
  productId: string;
  userId: string;
  userName: string;
  userPhoto?: string;
  rating: number;
  comment: string;
  images?: string[];
  replies?: any[];
  createdAt: number;
}